Changed import structure to support tests
